alerts: fix support for multiple music alarms
When setting multiple music alarms at short intervals, there is a bug
in which the last sound source is played when the first alarm occurs.
so this was fixed.

This occurred because the AlertsAgent uses one audioplayer, and it was
fixed by modifying to have an audioplayer for each alarm.

Signed-off-by: Inho Oh <webispy@gmail.com>