fix ALBERTNUGU-1155
If set the timer at the same time after setting the alarm, the
ignore event is normally sent to the server when the timer occurs,
and a notification is displayed on the mobile app.

However, when setting the alarm at the same time after setting the
timer first, there is an issue that the notification is not
displayed in the mobile app because the deactivated timer is not
included in the context.

To fix the bug, the order of deactivation of items when an ignore
event occurs is changed to after event transmission, and logic
to collect items for 1 second to handle more than two ignores has
been added.

Signed-off-by: Inho Oh <webispy@gmail.com>